Simplified update and rollback for role and role_capability changes/

+++ traffic_ops/app/db/migrations/2023012316280200_create_new_role.down.sql
@@ -14,35 +14,9 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-DELETE FROM public.role_capability rc WHERE rc.cap_name='DELIVERY-SERVICE-SAFE:UPDATE' AND rc.role_id IN (SELECT id FROM public.role r WHERE r.priv_level < 20);
 
-INSERT INTO public.role_capability (role_id, cap_name)
-SELECT id, perm
-FROM public.role
-         CROSS JOIN ( VALUES
-                          ('DELIVERY-SERVICE:UPDATE')
-) AS perms(perm)
-WHERE "priv_level" < 20 AND "priv_level" > 0
+DELETE FROM public.role_capability rc WHERE  rc.role_id='trouter'
     ON CONFLICT DO NOTHING;
 
-
-
-
-DELETE FROM public.role_capability
-WHERE role_id=236 AND cap_name='CDN:READ';
-DELETE FROM public.role_capability
-WHERE role_id=236 AND cap_name='DELIVERY-SERVICE:READ';
-DELETE FROM public.role_capability
-WHERE role_id=236 AND cap_name='DNS-SEC:READ';
-DELETE FROM public.role_capability
-WHERE role_id=236 AND cap_name='FEDERATION:READ';
-DELETE FROM public.role_capability
-WHERE role_id=236 AND cap_name='STEERING:READ';
-DELETE FROM public.role_capability
-WHERE role_id=236 AND cap_name='FEDERATION-RESOLVER:READ';
-DELETE FROM public.role_capability
-WHERE role_id=236 AND cap_name='DS-SECURITY-KEY:READ';
-
-DELETE FROM public."role"
-WHERE id=236;
-
+DELETE FROM public."role" r WHERE r.name='trouter'
+    ON CONFLICT DO NOTHING;
